0_exif_data/requirements.txt
piexif
ExifRead
Pillow
